Added definition of references in the snapshot collection.
Incorporated sorting difference data to the demo tests. This is needed when the natural sort order of the database data is variable - in this case due to GUID primary keys, which means the data is returned in a random order depending on the ids that happened to be allocated.

This is actually quite a good proof of the whole concept of unpredictable data. The output contains unpredictable keys and timestamps, and yet a consistent difference report is generated in each run.